Move capi types to CAPI API
