+++ public/client-v2/src/components/CAPI/FrontsTagInput.js
