--- public/client-v2/src/components/PandaProvider.js
@@ -1,6 +1,0 @@
-// @flow
-
-import * as React from 'react';
-
-// $FlowFixMe: waiting for React to update flow types
-const PandaContext = React.createContext();
