+++ public/client-v2/src/services/capiQuery.js
@@ -6,11 +6,54 @@ const API_BASE = 'https://content.guardianapis.com/';
 
 type Fetch = (path: string) => Promise<Response>;
 
+type ImageAsset = {
+  type: 'image',
+  mimeType: string,
+  file: string,
+  typeData: {
+    width: string,
+    number: string
+  }
+};
+
+type ImageElement = {
+  id: string,
+  relation: string,
+  type: 'image',
+  assets: ImageAsset[]
+};
+
+type Element = ImageElement;
+
+type Article = {
+  webTitle: string,
+  webUrl: string,
+  elements?: Element[]
+};
+
+type CAPISearchQueryReponse = {
+  response: {
+    results: Article[]
+  }
+};
+
+type Tag = {
+  id: string,
+  webTitle: string,
+  webUrl: string
+};
+
+type CAPITagQueryReponse = {
+  response: {
+    results: Tag[]
+  }
+};
+
 const capiQuery = (
   baseURL: string = API_BASE,
   fetch: Fetch = window.fetch
 ) => ({
-  search: async (params: Object) => {
+  search: async (params: Object): Promise<CAPISearchQueryReponse> => {
     const response = await fetch(
       `${baseURL}search${qs({
         ...params
@@ -19,7 +62,7 @@ const capiQuery = (
 
     return response.json();
   },
-  tags: async (params: Object) => {
+  tags: async (params: Object): Promise<CAPITagQueryReponse> => {
     const response = await fetch(
       `${baseURL}tags${qs({
         ...params
@@ -30,5 +73,5 @@ const capiQuery = (
   }
 });
 
-export type { Fetch };
+export type { Fetch, Element };
 export default capiQuery;
